requirements.txt
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
filelock==3.19.1
fsspec==2025.7.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pycparser==2.22
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sounddevice==0.5.2
srt==3.5.3
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.15.0
urllib3==2.5.0
vosk==0.3.44
websockets==15.0.1
